+++ CHANGELOG.md
@@ -5,5 +5,6 @@ Simple SMS
 
 ####Alpha 2
 * Added [EZTexting Driver](https://www.eztexting.com/)
+* Added [CallFire Driver](https://www.callfire.com/)
 * Fixed a bug where the `pretend` configuration variable was not working.
 * The E-Mail Driver will now throw an error if a carrier is not found or provided.